Merge in SECPAR/hammer from bugfix/sign-conversions to release/v3.0.0

* commit '24f5a2dcac0b9f22622b04a9a813aff9182b7bc9':
  I'm doing this as a test because I'm having trouble connecting to bitbucket
  Fixed sign conversions that could potentially cause overflows, left sign conversions in bits.c due to them being on purpose

+++ src/hammer.h
@@ -163,7 +163,7 @@ typedef struct HParsedToken_ {
  */
 typedef struct HParseResult_ {
     const HParsedToken *ast; /**< Abstract syntax tree */
-    int64_t bit_length;
+    size_t bit_length;
     HArena *arena; /**< Memory arena for the parse result */
 } HParseResult;
